requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
astroid==3.3.4
certifi==2024.8.30
click==8.1.7
coverage==7.6.1
dill==0.3.9
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==24.1
pep8==1.7.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pylint==3.3.1
pytest==8.3.3
pytest-cov==5.0.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tomli==2.0.1
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.31.0
